refactored fields
