+++ src/Features/Admin/Users/EditUser/index.js
@@ -4,25 +4,21 @@ import CommonGridBasedForm from '../../../../components/CommonGridBasedForm';
 import { SELECT, TEXT_FIELD } from '../../../../components/CommonGridBasedForm/FieldTypes';
 import { emailRegex } from '../../../../redux/ActionTypes';
 import { contactRegex } from '../../../../scripts/constants';
-import { validateOnSubmit } from '../../../../util/FieldsValidCheckOnForm';
+import { validateOnSubmit, fieldChangeHandler } from '../../../../util/CommonGridBasedFormUtils';
 
 const AddUser = () => {
   const [onSaveSuccess, setOnSaveSuccess] = useState(false);
-  const [role, setRole] = useState([]);
-  const [email, setEmail] = useState('');
-  const [password, setPassword] = useState('');
-  const [contact, setContact] = useState('');
   const [fields, setFields] = useState([
     {
       type: SELECT,
       label: 'Role',
       values: ['User', 'Vendor'],
-      value: role,
+      value: [],
       errorMessage: '',
 
       onChange: ({ target: { value } }, index) => {
-        setRole(value);
-        fields[index].value = value;
+        const updatedFields = fieldChangeHandler(fields, value, index);
+        setFields(updatedFields);
       },
     },
     {
@@ -30,19 +26,17 @@ const AddUser = () => {
       textFieldType: 'email',
       label: 'Email',
       variant: 'standard',
-      value: email,
+      value: '',
       errorMessage: '',
       onChange: ({ target: { value } }, index) => {
-        setEmail(value);
-        fields[index].value = value;
-        fields[index].getValidation(value, index);
+        const updatedFields = fieldChangeHandler(fields, value, index);
+        setFields(updatedFields);
       },
-      getValidation: (value, index) => {
+      getValidation: (value) => {
         if (!emailRegex.test(value)) {
-          fields[index].errorMessage = 'Email type is not valid';
-        } else {
-          fields[index].errorMessage = '';
+          return 'Email type is not valid';
         }
+        return '';
       },
     },
     {
@@ -50,19 +44,17 @@ const AddUser = () => {
       textFieldType: 'password',
       label: 'Password',
       variant: 'standard',
-      value: password,
+      value: '',
       errorMessage: '',
       onChange: ({ target: { value } }, index) => {
-        setPassword(value);
-        fields[index].value = value;
-        fields[index].getValidation(value, index);
+        const updatedFields = fieldChangeHandler(fields, value, index);
+        setFields(updatedFields);
       },
-      getValidation: (value, index) => {
+      getValidation: (value) => {
         if (value.length < 8) {
-          fields[index].errorMessage = 'Password must be 8 characters long';
-        } else {
-          fields[index].errorMessage = '';
+          return 'Password must be 8 characters long';
         }
+        return '';
       },
     },
     {
@@ -70,19 +62,17 @@ const AddUser = () => {
       textFieldType: 'text',
       label: 'Contact',
       variant: 'standard',
-      value: contact,
+      value: '',
       errorMessage: '',
       onChange: ({ target: { value } }, index) => {
-        setContact(value);
-        fields[index].value = value;
-        fields[index].getValidation(value, index);
+        const updatedFields = fieldChangeHandler(fields, value, index);
+        setFields(updatedFields);
       },
-      getValidation: (value, index) => {
+      getValidation: (value) => {
         if (!contactRegex.test(value)) {
-          fields[index].errorMessage = 'Contact length or Type is not valid';
-        } else {
-          fields[index].errorMessage = '';
+          return 'Contact length or Type is not valid';
         }
+        return '';
       },
     },
   ]);
